Update geom_range.R to fix issue #306
Change the signs (+ and -) in order to correctly plot the lower and upper values of the ranges in geom_range.
Before this change, the values were "mirrored", leading to incorrect range displays when the heights of the nodes were not exactly at the center of the range. Now, the values are "transposed", which corrects this issue.
This works both with or without adding the revts() function.